CI: no longer pin UNINATIVE_VERSION to 3.9
There was a bug in Uninative 4.0 which came with YP 3.1.27 that would break
CI, so it had to be pinned down to version 3,9 until resolved upstream. Now
YP 3.1.28 comes with Uninative 4.3 with the fix, so now longer needs to be
pinned down.

This reverts commit 1d23a215a4dc5ec77c3e0b25f86aabdb252285d4.

Bug-AGL: SPEC-4875

Change-Id: I8b7789630179895ba7b82b3d99625459e3664a78
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/29259
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>