meta-agl-bsp: fixes for meta-rcar-gen3 changes
The scarthgap-dev branch in meta-renesas now contains a complete set
of gstreamer 1.20.3 components that are set as the preferred versions
by default when the "multimedia" machine feature is set, which AGL
does for the rcar-gen3 machines.  To accommodate this, remove our
BBMASKing of meta-rcar-gen3 gstreamer recipes, and add back the
rcar-gen3 version of the patch to gstreamer1.0-plugins-bad.

Bug-AGL: SPEC-5239

Change-Id: I55aecbf6d9a32519ec5f6f2e1b895fb6f5087b71
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/30205
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>