wireplumber: fix audio randomly not working at boot time
This issue is caused by a race condition that triggers some links to be
created and then destroyed very fast, but at a certain point where the
managing object (si-standard-link) is not fully configured yet. At that
point, droping the ref count of the si-standard-link doesn't destroy it
and the links remain dangling, causing a loop in the graph that prevents
audio from working properly.

This patch mitigates the issue by forcing si-standard-link to reset to
its initial state, forcing the pipewire links to be destroyed. This does
not, however, prevent the potential leak of the si-standard-link.
A better solution is being worked on for upstreaming.

Bug-AGL: SPEC-5244
Change-Id: I61477c0c021351903239d58fff0217e25e08db02
Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/30658
ci-image-boot-test: Jenkins Job builder account
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>