meta-agl-flutter: Add flutter-auto bbappend
Add a flutter-auto bbappend to apply a local patch to use pkgconfig
for the EGL and GLESv2 library link options.  This fixes building
with latest meta-rockchip while still being compatible with other
BSPs.  When upstream takes this or a similar fix and we get it on
a meta-flutter upgrade, this can be removed.

Bug-AGL: SPEC-5438

Change-Id: I9a2ca4e902b239cacf2f38ccf9fce37ad19b27e2
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/31053
Reviewed-by: Naoto YAMAGUCHI <naoto.yamaguchi@aisin.co.jp>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
ci-image-boot-test: Jenkins Job builder account