meta-agl-kuksa-val: Fix kuksa-client build
Building the kuksa-client wheel is hitting an odd dependency
failure that does not seem sensible.  For now, pass the dependency
skipping option to the wheel build to avoid this.  This should be
revisited when setuptools is next upgraded in upstream.

Also remove the grpcio-tools dependency tweak, as meta-python now has
a version that meets the requirement.

Bug-AGL: SPEC-5147

Change-Id: I7d73be022cf0a4f8027c81bffba27837a8a96fa8
Signed-off-by: Scott Murray <scott.murray@konsulko.com>