meta-agl-kuksa-val: upgrade kuksa-client
Upgrade kuksa-client (aka kuksa-python-sdk) to its latest commit to
pick up all the VAL API v2 changes since the 0.4.3 release.  Upstream
have tagged a couple of 0.5.0 alpha releases, so the latest of those
(0.5.0-a2) has been used as the base version number for now.

In discussion with upstream, a final 0.5.0 release does not seem
imminent, but they consider the current tree to be stable for VAL API
v1 support.  Since the v1 API is now considered deprecated in the
databroker, taking an upgrade now to allow testing with the v2 API
seems justified.  Initial testing did find an oversight with respect
to v1 versus v2 behavior that a new local patch fixes, it will be
submitted upstream ASAP.

Bug-AGL: SPEC-5453

Change-Id: I44b7cb306d3773f95fe85b5f769ec5e2d9e967a9
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/31084
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>