Changes for RPI500
The PI500 is not yet fully supported in meta-raspberrypi (scarthgap).
Add a few tweaks to allow it to boot.
For simplicity, we add a full machine template 'raspberrypi500'.
It is not yet possible to do it with the same configuration as the pi5.

We need to skip U-boot (use pi bootloader) and disable serial.

Bug-AGL: SPEC-5477

Change-Id: I47ded55b870a4d69cf6347f4407c1ab9c9f71b4d
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/31161
ci-image-build: Jenkins Job builder account
Tested-by: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account