meta-agl-bsp: Force 6.12 kernel for RPi
meta-raspberrypi has switched back to the 6.6 kernel by default,
as the switch to 6.12 was deemed too big of a change for a LTS
branch.  We, however, do want to use the newer kernel to pick up
fixes and new features, so set PREFERRED_VERSION to force it.

Bug-AGL: SPEC-5480

Change-Id: I4f13bab690ab37b5cd750ee8c3ddf517bff323f3
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/31174
ci-image-build: Jenkins Job builder account
Reviewed-by: Marius Vlad <marius.vlad@collabora.com>
ci-image-boot-test: Jenkins Job builder account
Tested-by: Jenkins Job builder account