meta-agl-core: remove SystemTap support
ATM the Sparrow Hawk GPU driver source has function tracing support that
fails to build due to a mismatch between the driver and the Sparrow Hawk
BSP's 6.12 kernel, due to a tracing change that was introduced in 6.10.
This becomes a problem when "agl-devel" is enabled, as we enable function
tracing support explicitly for potential SystemTap use.  Since we do not
have any in-tree users of SystemTap or ftrace, it seems reasonable to
replace this with just enabling minimal perf tracing support to avoid
the issue for now.  If a downstream needs function tracing support, they
can still enable it.

Bug-AGL: SPEC-5544

Change-Id: Ie6210079db41cab59f8e88832b599e7bcc1d8bb6
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/31364
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Naoto YAMAGUCHI <naoto.yamaguchi@aisin.co.jp>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account